[MOVE] : DI 패키지 구조 변경 extension 외부로 이동
