+++ app/src/main/java/com/indipage/di/Qualifier.kt
@@ -1,4 +1,4 @@
-package com.indipage.di.extension
+package com.indipage.di
 
 import javax.inject.Qualifier
